+++ app/src/main/java/com/canopas/yourspace/ui/flow/journey/components/LocationHistory.kt
@@ -1,4 +1,4 @@
-package com.canopas.yourspace.ui.flow.home.map.member.components
+package com.canopas.yourspace.ui.flow.journey.components
 
 import android.location.Address
 import androidx.compose.foundation.Image
